Fix iso9660.h SSIZE_T type on non-MSVC builds
SSIZE_T (uppercase) is a Windows-only typedef pulled in via
<windows.h>, which the header only includes under _MSC_VER. On Linux
the bytes-remaining counter inside extractFile() falls back to an
undeclared SSIZE_T and the build fails. ISO 9660 sizes are 32-bit
anyway, so just use uint32_t.

Without this, archive.cpp (and therefore both the standalone nuance
binary and nuance_libretro.so) fails to compile on Linux.

+++ src/iso9660.h
@@ -165,9 +165,9 @@ public:
 
         ISO_FSEEK(fp, (long long)lba * 2048, SEEK_SET);
         uint8_t buf[65536];
-        SSIZE_T remaining = size;
+        uint32_t remaining = size;
         while (remaining > 0) {
-            SSIZE_T chunk = remaining < sizeof(buf) ? remaining : sizeof(buf);
+            uint32_t chunk = remaining < sizeof(buf) ? remaining : sizeof(buf);
             size_t rd = fread(buf, 1, chunk, fp);
             if (rd == 0) break;
             fwrite(buf, 1, rd, out);
